Improved output pad reuse.
Added new method for getting an output pad for tools.
Instead of having an output pad specific for each tool, now all can
share the same pad.

Also added an Id argument to GetOutputProgressMonitor. This id
is used when building an Id for the corresponding key binding.
Until now, we were using the monitor title for this, but it doesn't
work well since the title is localized.