Work around GTK context menu bug
Triggering context menu from a button press event (which is the
correct thing to do) on mac is problematic because the release
event sometimes causes the menu to hide again (e.g. sln pad but
only when docked on left side). Work around this by triggering
it on the release event instead.