Remove break event request from breakpoints dictionary when removing a breakpoint. Fixes issue with removed breakpoint being added again in HandleAssemblyUnloadEvents.
When adding and removing a breakpoint while the soft debugger is attached, the breakpoint is not removed from the breakpoints dictionary.
If the assembly is unloaded after the add and remove of a breakpoint, then HandleAssemblyUnloadEvents will add the removed breakpoint from breakpoints dictionary to pending_bes, causing the removed breakpoint to be added again.