[Debugger] Fix incorrect ordering for evaluated group in Locals window
When an evaluating group's ObjectValue is updated the items were
added in reverse order in the Locals window. The first item was
replacing the existing item but then the rest were inserted one at
a time after the first item. This resulted in all items after the
first one being added in reverse order.