Record evaluation duration
The AsyncEvaluationTracker now times how long an evaluation takes and
records the information on the DebuggerSession's EvaluationStatistics
class. The EvaluationStatistics class records the min, max and average
evaluation times in milliseconds, as well as the number of successful
evaluations and the number of failed evaluations.

No current support capturing evaluation times for individual items
when a group is being evaluated. EvaluatingGroup may be passed to
the AsyncEvaluationTracker if there is no FrameInfo available.