Reduce amount of code to record evaluation duration
Added an EvaluationTimer class that handles the timing for how long
the evaluation takes and recording the information with the
EvaluationStatistics class. The EvaluationTimer class can be created
from the EvaluationStatistics class from its StartTimer method.

Also correctly identify evaluation failures. Previously only if the
evaluation threw an exception would a failure be indicated. Now
the ObjectValue itself has its Flags checked to see if the evaluation
was successful. If there is a failure an ObjectValue will be created
but will have a flag that can be used to detect if the evaluation
was successful or not.