Record evaluation duration when evaluating groups
If the AsyncEvaluationTracker is evaluating when ObjectValues are
requested from the BaseBacktrace the evaluation is done as a group
so the timing from the AsyncEvaluationTracker cannot be used to find
individual evaluation times. To capture these individual evaluation
times a timer is now used in the BaseBacktrace class when getting
parameters, local variables, exceptions and the 'this' reference
if these are being evaluated as a group.