SoftDebuggerAdaptor: support CoreFX Nullable implementation
When Mono switched to the Nullable implementation from CoreFX with https://github.com/mono/mono/commit/275d7bfc2ca17aa03a9daeb32493ee0ef9078037
some tests in VSMac broke because the SoftDebuggerAdaptor is checking for the old field name in the type.

Updated the adaptor to support both the CoreFX and the old Mono field name.