Added EvaluationContext to AsyncOperation
SoftDebugger's MethodCall async op already had it (Cor did not),
but we need this to report telemetry on what evaluation options
were enabled for the particular async op that is busy.

Partial "fix" to hopefully allow us to get more info about why/when
this is happening for our users.

https://devdiv.visualstudio.com/DevDiv/_workitems/edit/802365/