Creating a new version of MOD_KIND_EXCEPTION_ONLY to support Not Filtered Exception feature, discussed with Joaquin Jares. (#257)
Because we don't have this functionality of get every other exception that is not specified, VsWin was getting every exception and filtering later. But this causes a slowness during the debugger because we stop all thread on each exception that we get, even if they are caught and we don't want to stop.

This fix a bug of slowness on Android Debugger.