Only resume VM when suspended.
Running VM with suspend="n" (for late attach scenarios), raise a number
of errors in debugger client. This is due to runtime not being suspended
for several events, but debugger client didn't check the suspend policy
returned by VM and always tried to do a resume triggering a
VMNotSuspendedException for several different event types.

Fix is to check the returned suspend policy (since that indicates if VM
is in a suspend state or not) and only resume when suspended.