Use a different lock to manage breakpointStore
Fixes VSTS 1026591
SigTerm signal in Mono.Debugging.dll!Mono.Debugging.Client.DebuggerSession::get_Breakpoints+9

`slock` is used throughout DebuggerSession, however in this case we only need to guard against
multiple threads assigning an instance of BreakpointStore. Also, since BreakpointStore is thread-safe
we can remove the additional locks around it where we lock on enumeration of `breakpointStore`.

This fixes the case where one thread dispatches an OnExit (for example) which tries communicate
with the debuggee and waits on data from the socket and another thread attempts to clean up breakpoints.