[Mono.Debugging.Soft] Simplified breakpoint resolving logic
The old logic was trying to be so smart, it smartness-overloaded
into Smartness.MinValue.

In all seriousness, though, now that the IDE intelligently resolves
breakpoints to the correct line, we don't need to be "smart".

This code was only needed 10 years ago when the IDE would put
breakpoints on blank lines and so the debugger would need to
do insanely fuzzy matching. So fuzzy it needed glasses, apparently.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1094719/