Proper support for Catchpoint location signature
When an exception happens we didn't correctly retrieve and store the location of the exception.

We used to append the module name even if the method name already contained the module name (so we'd end up with Module!Module.dll!Class.Method). Also the Address is 0 most of the time for VSCode Debug Protocol, so no need to append it if it's zero.

Finally, store the location signature on the Catchpoint itself, so that when the user selects "Ignore exceptions at this location" we use the exact location that the debugger gave us and add it to the ignore list. Otherwise the exception can be on a different frame (the current frame that has the source code, and not the bottom frame of the stack).