Retrieve only necessary assembly info for debug session (#413)
As we are reading all the assembly information from protocol, scenarios where the debugger goes over the network are heavily impacted by the amount of information we get.

By not retrieving information that's not necessary we are improving the assembly load times, which improves the app launch times. This is specially important for remote iOS scenarios from Visual Studio.