Use ConcurrentDictionary for storing reply packets and callbacks, to relax explicit lock usage
A PerfWatson issue shows a high amount of hits caused by the excessive locking when sending and receiving packets, that causes perf issues especially if a packet takes time to be received: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2400681

Using ConcurrentDictionary to store the packets and callbacks, let the dictionary implementation to handle the different access locks and ensure performance and efficiency.
We still maintain the lock object but only for disconnection checks.